+++ front/pages/api/v1/w/[wId]/apps/index.ts
@@ -1,13 +1,4 @@
-import type { AppType, WithAPIErrorResponse } from "@dust-tt/types";
-import type { NextApiRequest, NextApiResponse } from "next";
-
-import { Authenticator, getAPIKey } from "@app/lib/auth";
-import { AppResource } from "@app/lib/resources/app_resource";
-import { apiError, withLogging } from "@app/logger/withlogging";
-
-export type GetAppsResponseBody = {
-  apps: AppType[];
-};
+import handler from "@app/pages/api/v1/w/[wId]/vaults/[vId]/apps";
 
 /**
  * @swagger
@@ -78,37 +69,4 @@ export type GetAppsResponseBody = {
  *         description: Internal Server Error.
  */
 
-async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<GetAppsResponseBody>>
-): Promise<void> {
-  const keyRes = await getAPIKey(req);
-  if (keyRes.isErr()) {
-    return apiError(req, res, keyRes.error);
-  }
-  const { workspaceAuth } = await Authenticator.fromKey(
-    keyRes.value,
-    req.query.wId as string
-  );
-
-  switch (req.method) {
-    case "GET":
-      res.status(200).json({
-        apps: (await AppResource.listByWorkspace(workspaceAuth)).map((app) =>
-          app.toJSON()
-        ),
-      });
-      return;
-
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
-        },
-      });
-  }
-}
-
-export default withLogging(handler);
+export default handler;
